+++ src/lib/input_handling/InputComponent.ts
@@ -6,31 +6,39 @@ export default class InputComponent {
   requiresFocus: boolean;
   input: Input;
   disabled: boolean = false;
-  onpressedCallback: (() => boolean) | undefined;
-  onholdCallback: (() => void) | undefined;
-  onreleaseCallback: (() => void) | undefined;
+  onpressedCallback?: () => void;
+  onholdCallback?: () => void;
+  onreleaseCallback?: () => void;
+  consumePress: boolean;
 
   constructor(
       input: Input,
       focusElement?: HTMLElement,
       requiresFocus: boolean = false,
-      onpressed: (() => boolean) | undefined = undefined,
-      onhold: (() => void) | undefined = undefined,
-      onrelease: (() => void) | undefined = undefined) {
+      onpressed?: () => void,
+      onhold?: () => void,
+      onrelease?: () => void,
+      consumePress: boolean = false) {
     this.onpressedCallback = onpressed;
     this.onholdCallback = onhold;
     this.onreleaseCallback = onrelease;
     this.focusElement = focusElement;
     this.input = input;
-    console.log(focusElement);
     this.requiresFocus = requiresFocus;
+    this.consumePress = consumePress;
   }
 
   onpressed(): boolean {
-    if (this.onpressedCallback) {
-      return this.onpressedCallback();
+    if (this.disabled) {
+      // forward to next button.
+      return false;
     }
-    return false
+    if (this.onpressedCallback) {
+      this.onpressedCallback();
+    }
+    // We normally don't consume this pressed event, 
+    // so other components can react to the same press at the same time.
+    return this.consumePress;
   }
 
   onhold() {
@@ -69,22 +77,26 @@ export default class InputComponent {
   }
 
   // --- Keyboard ---
-  thisKey(event?: KeyboardEvent): boolean {
+  thisKey(event: KeyboardEvent): boolean {
     return false;
   }
 
   // key pressed (down)
-  onkeypressed(event?: KeyboardEvent): boolean {
-    return this.onpressed();
+  onkeypressed(event: KeyboardEvent): boolean {
+    return this.thisKey(event) && this.onpressed();
   }
 
   // key released (up)
-  onkeyrelease(event?: KeyboardEvent) {
-    this.onrelease();
+  onkeyrelease(event: KeyboardEvent) {
+    if (this.thisKey(event)) {
+      this.onrelease();
+    }
   }
 
   // key hold down
-  onkeyhold(event?: KeyboardEvent) {
-    this.onhold();
+  onkeyhold(event: KeyboardEvent) {
+    if (this.thisKey(event)) {
+      this.onhold();
+    }
   }
 }