+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.0-rc4
+## API-changes
+ - consumePress-flag that determines if button should consume events and stop forwarding to other components (default behavior is false)
+ - onclick-callback is void instead of boolean (onclick can not interfere bubbling behavior of press events anymore)
+
 # 1.0.0-rc3
 ## API-changes
  - update exports, include (un)registerComponent